requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai>=1.40.0
python-dotenv>=1.0.1